+++ api/web/public/index.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="en" class='h-full'>
     <head>
         <meta charset="utf-8">
         <meta http-equiv="X-UA-Compatible" content="IE=edge">
@@ -10,10 +10,10 @@
         <link type='text/css' href='/assembly.min.css' rel='stylesheet'/>
         <script async defer src='/assembly.js'></script>
     </head>
-    <body>
+    <body class='h-full'>
         <noscript>
             <strong>We're sorry but <%= htmlWebpackPlugin.options.title %> doesn't work properly without JavaScript enabled. Please enable it to continue.</strong>
         </noscript>
-        <div id="app"></div>
+        <div id="app" class='h-full'></div>
     </body>
 </html>
