+++ api/web/src/pages/map/main.js
@@ -2,6 +2,7 @@ import Vue from 'vue'
 import VueRouter from 'vue-router';
 import VTooltip from 'v-tooltip';
 import App from './App.vue'
+import std from '../../std.js';
 
 Vue.config.productionTip = false
 Vue.use(VueRouter);
